+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Upcoming
 - Improved query performance as the number of components grows.  This actually reduced performance on some of the synthetic benchmarks but should have a greatly positive effect in actual applications.
+- Improved performance of component initialization, both with and without custom field values.
 - Documented query builder and query object.
 
 ### 0.8.0